quick commit

+++ L-Galaxies_development/code/allvars.c
@@ -251,7 +251,7 @@ float FilterLambda[NMAG];	//wavelength of each filter
 #ifdef SPEC_PHOTABLES_ON_THE_FLY
 int NLambdaFilter[NMAG];
 #ifdef REIONIZEPHOTON
-float NPhotTables[SSP_NMETALLICITES][SSP_NAGES];
+double NPhotTables[SSP_NMETALLICITES][SSP_NAGES];
 #endif
 #endif
